Added cbot_common package for all common functionalities
